cpq/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
email-validator==2.0.0

